+++ parivesh/src/main/java/com/backend/repository/postgres/ThresholdRepository.java
@@ -12,11 +12,11 @@ public interface ThresholdRepository extends JpaRepository<ThresholdParameter, I
 
 	
 	@Query("SELECT new ThresholdParameter(t.id, t.name,t.unit,t.val,t.threshold_unit,t.rendering_type,t.regex,t.description) from ThresholdParameter t where t.is_active='true' and t.is_deleted='false' and t.activity_id=?1")
-    List<ThresholdParameter> findThresholdbyActId(Long id);
+    List<ThresholdParameter> findThresholdbyActId(Integer id);
     
     @Query("SELECT new ThresholdParameter(t.id, t.name,t.unit,t.val,t.threshold_unit,t.rendering_type,t.regex,t.description) from ThresholdParameter t where t.is_active='true' and t.is_deleted='false' and t.subactivity_id=?1")
-    List<ThresholdParameter> findThresholdbySubActId(Long id);
+    List<ThresholdParameter> findThresholdbySubActId(Integer id);
     
     @Query("SELECT new ThresholdParameter(t.id, t.name,t.unit,t.val,t.threshold_unit,t.rendering_type,t.regex,t.description) from ThresholdParameter t where t.is_active='true' and t.is_deleted='false' and t.activity_id=?2 and t.subactivity_id=?1")
-    List<ThresholdParameter> findThreshold(Long ActId,Long SubId);
+    List<ThresholdParameter> findThreshold(Integer ActId,Integer SubId);
 }
